requirements.txt
aiogram==2.18
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.8.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.10
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
pycoingecko==2.2.0
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
yarl==1.7.2
